# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Booking Service",
    "description": "Микросервис для бронирования отелей: отели, комнаты, пользователи и бронирования",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api/hotel-booking-service"
    }
  ],
  "paths": {
...
    "/api/v1/bookings/": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Create Booking",
        "description": "Создание бронирования",
        "operationId": "create_booking_api_v1_bookings__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_BookingRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Bookings"],
        "summary": "List Bookings",
        "description": "Реестр бронирований",
        "operationId": "list_bookings_api_v1_bookings__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_List_BookingWithDetailsRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/{booking_id}": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get Booking",
        "operationId": "get_booking_api_v1_bookings__booking_id__get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_BookingWithDetailsRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Bookings"],
        "summary": "Update Booking",
        "operationId": "update_booking_api_v1_bookings__booking_id__put",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_BookingRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Bookings"],
        "summary": "Delete Booking",
        "operationId": "delete_booking_api_v1_bookings__booking_id__delete",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookingCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "check_in_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check In Date"
          },
          "check_out_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check Out Date"
          },
          "number_of_guests": {
            "type": "integer",
            "title": "Number Of Guests"
          },
          "special_requests": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Requests"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "hotel_id",
          "room_id",
          "check_in_date",
          "check_out_date",
          "number_of_guests"
        ],
        "title": "BookingCreate",
        "description": "Схема для создания бронирования"
      },
      "BookingRead": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "check_in_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check In Date"
          },
          "check_out_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check Out Date"
          },
          "number_of_guests": {
            "type": "integer",
            "title": "Number Of Guests"
          },
          "special_requests": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Requests"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "booking_reference": {
            "type": "string",
            "title": "Booking Reference"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "booking_reference",
          "total_price",
          "created_at"
        ],
        "title": "BookingRead"
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "confirmed",
          "cancelled",
          "completed",
          "checked_in",
          "checked_out"
        ],
        "title": "BookingStatus"
      },
      "BookingUpdate": {
        "properties": {
          "check_in_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check In Date"
          },
          "check_out_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Out Date"
          },
          "number_of_guests": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Guests"
          },
          "special_requests": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Requests"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BookingUpdate"
      },
      "BookingWithDetailsRead": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "check_in_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check In Date"
          },
          "check_out_date": {
            "type": "string",
            "format": "date-time",
            "title": "Check Out Date"
          },
          "number_of_guests": {
            "type": "integer",
            "title": "Number Of Guests"
          },
          "special_requests": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Requests"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "booking_reference": {
            "type": "string",
            "title": "Booking Reference"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          },
          "hotel": {
            "$ref": "#/components/schemas/HotelRead"
          },
          "room": {
            "$ref": "#/components/schemas/RoomRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "booking_reference",
          "total_price",
          "created_at"
        ],
        "title": "BookingWithDetailsRead"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Информация об ошибке"
      },
...
      "HotelRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "HotelRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoomRead": {
        "properties": {
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_night": {
            "type": "number",
            "title": "Price Per Night"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "amenities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hotel_id",
          "created_at"
        ],
        "title": "RoomRead"
      },
      "RoomStatus": {
        "type": "string",
        "enum": [
          "available",
          "occupied",
          "maintenance",
          "cleaning"
        ],
        "title": "RoomStatus"
      },
...
      "SuccessResponse_BookingRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/BookingRead"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[BookingRead]"
      },
      "SuccessResponse_BookingWithDetailsRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/BookingWithDetailsRead"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[BookingWithDetailsRead]"
      },
...
      "SuccessResponse_List_BookingWithDetailsRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/BookingWithDetailsRead"
            },
            "type": "array",
            "title": "Data"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[List[BookingWithDetailsRead]]"
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "UserRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}